refactor(controller): extract GPU sharding helpers to gpu_sharding.go
Part 2/7 of splitting the god-controller into per-concern files.
Tracks #311.

Move the multi-GPU sharding helpers — which translate the Model CRDs
GPUShardingSpec into llama.cpp --split-mode and --tensor-split flag
values — out of inferenceservice_controller.go into a focused
gpu_sharding.go sibling file.

Moved symbols:
  const ( splitModeLayer, splitModeRow, splitModeNone )
  resolveSplitMode
  calculateTensorSplit
  parseLayerRange
  gcd

Callers in the same package continue to resolve the unqualified
names; no callsite updates needed.

Also drops the now-unused "strconv" import from the controller file
(strconv was only used by parseLayerRange and calculateTensorSplit).

Line deltas (this PR is based on main; if C1 #312 merges first, the
controller file drops another 104 lines on rebase):
  inferenceservice_controller.go: 1,567 → 1,470 (−97)
  gpu_sharding.go:                        0 → 126 (new)

Verified: go vet clean, golangci-lint 0 issues, make test green,
internal/controller coverage unchanged at 83.1%.
